requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
-e file:///Users/mark/src/github.com/relaxdiego/media-server
    # via -r requirements.in
astroid==2.11.4
    # via pylint
attrs==21.4.0
    # via pytest
bcrypt==3.2.2
    # via paramiko
black==22.3.0
    # via -r requirements.in
certifi==2021.10.8
    # via requests
cffi==1.15.0
    # via
    #   bcrypt
    #   cryptography
    #   pynacl
cfgv==3.3.1
    # via pre-commit
charset-normalizer==2.0.12
    # via requests
click==8.1.3
    # via
    #   black
    #   pyinfra
colorama==0.4.4
    # via pyinfra
configparser==5.2.0
    # via pyinfra
cryptography==37.0.1
    # via
    #   paramiko
    #   requests-ntlm
dill==0.3.4
    # via pylint
distlib==0.3.4
    # via virtualenv
distro==1.7.0
    # via pyinfra
filelock==3.6.0
    # via virtualenv
flake8==4.0.1
    # via -r requirements.in
gevent==21.12.0
    # via pyinfra
greenlet==1.1.2
    # via gevent
identify==2.5.0
    # via pre-commit
idna==3.3
    # via requests
iniconfig==1.1.1
    # via pytest
isort==5.10.1
    # via pylint
jinja2==3.1.2
    # via pyinfra
lazy-object-proxy==1.7.1
    # via astroid
markupsafe==2.1.1
    # via jinja2
mccabe==0.6.1
    # via
    #   flake8
    #   pylint
mypy-extensions==0.4.3
    # via black
nodeenv==1.6.0
    # via pre-commit
ntlm-auth==1.5.0
    # via requests-ntlm
packaging==21.3
    # via pytest
paramiko==2.10.4
    # via pyinfra
pathspec==0.9.0
    # via black
pep8==1.7.1
    # via -r requirements.in
platformdirs==2.5.2
    # via
    #   black
    #   pylint
    #   virtualenv
pluggy==1.0.0
    # via pytest
pre-commit==2.18.1
    # via -r requirements.in
py==1.11.0
    # via pytest
pycodestyle==2.8.0
    # via flake8
pycparser==2.21
    # via cffi
pydantic==1.9.0
    # via -r requirements.in
pyflakes==2.4.0
    # via flake8
pyinfra==2.1
    # via -r requirements.in
pylint==2.13.8
    # via -r requirements.in
pynacl==1.5.0
    # via paramiko
pyparsing==3.0.8
    # via packaging
pytest==7.1.2
    # via
    #   -r requirements.in
    #   pytest-testinfra
pytest-testinfra==6.7.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via pyinfra
pywinrm==0.4.3
    # via pyinfra
pyyaml==6.0
    # via
    #   -r requirements.in
    #   pre-commit
requests==2.27.1
    # via
    #   -r requirements.in
    #   pywinrm
    #   requests-ntlm
requests-ntlm==1.1.0
    # via pywinrm
six==1.16.0
    # via
    #   paramiko
    #   python-dateutil
    #   pywinrm
    #   virtualenv
toml==0.10.2
    # via pre-commit
tomli==2.0.1
    # via
    #   black
    #   pylint
    #   pytest
typing-extensions==4.2.0
    # via
    #   astroid
    #   black
    #   pydantic
    #   pylint
urllib3==1.26.9
    # via requests
virtualenv==20.14.1
    # via pre-commit
wrapt==1.14.1
    # via astroid
xmltodict==0.12.0
    # via pywinrm
zope-event==4.5.0
    # via gevent
zope-interface==5.4.0
    # via gevent

# The following packages are considered to be unsafe in a requirements file:
# setuptools
